moved profitability filter in component
