+++ src/containers/ProductTracker/ProductTrackerFilterSection/index.scss
@@ -59,76 +59,6 @@
           width: 86px;
         }
 
-        &__profitability-preset {
-          border: none !important;
-          box-shadow: unset !important;
-          background: white;
-
-          .profitability-preset-btn {
-            color: #98aeca !important;
-            border: 1px solid #98aeca !important;
-            box-shadow: unset !important;
-            border-top-left-radius: 100px !important;
-            border-bottom-left-radius: 100px !important;
-            border-top-right-radius: 0;
-            border-bottom-right-radius: 0;
-            font-weight: 600 !important;
-            padding: 0 10px 0 15px;
-            height: 32px;
-          }
-
-          .ui.dropdown {
-            color: #98aeca !important;
-            border: 1px solid #98aeca !important;
-            box-shadow: unset !important;
-            border-left: none !important;
-            border-top-right-radius: 100px !important;
-            border-bottom-right-radius: 100px !important;
-            border-top-left-radius: 0;
-            border-bottom-left-radius: 0;
-            font-weight: 600 !important;
-            padding: 10px 5px;
-            height: 32px;
-
-            .menu {
-              padding: 15px 0;
-              border-radius: 15px !important;
-              width: 204px;
-
-              .item:hover {
-                background-color: #98aeca !important;
-              }
-            }
-
-            .angle.down {
-              margin: 0 0.1em 0 -0.114286em !important;
-            }
-          }
-
-          &.basic {
-            .profitability-preset-btn {
-              color: #3b4557 !important;
-            }
-
-            .ui.dropdown {
-              color: #3b4557 !important;
-            }
-          }
-
-          &.blue {
-            .profitability-preset-btn {
-              color: #fff !important;
-              background-color: #349af8 !important;
-              border-right: 1px solid #4285f461 !important;
-            }
-
-            .ui.dropdown {
-              color: #fff !important;
-              background-color: #349af8 !important;
-            }
-          }
-        }
-
         &--period {
           width: 124px;
         }
